+++ src/components/molecules/ZopaFooter/ZopaFooter.tsx
@@ -44,6 +44,7 @@ const ZopaFooter = ({
         <FlexCol xs={12} s={6} l={3} className="mb-8 s:mb-9">
           <Heading className="mb-4 s:mb-6">What we do</Heading>
           <List>
+            <li className="mb-4">{renderLink({ href: `${baseUrl}/car-finance`, children: 'Car hire purchase' })}</li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/loans/car-loans`, children: 'Car loans' })}</li>
             <li className="mb-4">
               {renderLink({ href: `${baseUrl}/loans/debt-consolidation`, children: 'Debt consolidation loans' })}
@@ -52,11 +53,8 @@ const ZopaFooter = ({
               {renderLink({ href: `${baseUrl}/loans/home-improvement`, children: 'Home improvement loans' })}
             </li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/loans/wedding`, children: 'Wedding loans' })}</li>
-            <li className="mb-4">{renderLink({ href: `${baseUrl}/invest`, children: 'Peer-to-peer investments' })}</li>
-            <li className="mb-4">
-              {renderLink({ href: `${baseUrl}/invest/isa`, children: 'Innovative Finance ISA' })}
-            </li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/credit-card`, children: 'Credit cards' })}</li>
+            <li className="mb-4">{renderLink({ href: `${baseUrl}/smart-saver`, children: 'Smart Saver' })}</li>
             <li>{renderLink({ href: `${baseUrl}/savings-accounts`, children: 'Fixed Term Savings' })}</li>
           </List>
         </FlexCol>
@@ -72,7 +70,9 @@ const ZopaFooter = ({
             <li className="mb-4">{renderLink({ href: `${baseUrl}/about/awards`, children: 'Awards' })}</li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/about/careers`, children: 'Careers' })}</li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/blog`, children: 'Blog' })}</li>
-            <li>{renderLink({ href: `${baseUrl}/contact/complaints`, children: 'Complaints' })}</li>
+            <li className="mb-4">{renderLink({ href: `${baseUrl}/contact/complaints`, children: 'Complaints' })}</li>
+            <li className="mb-4">{renderLink({ href: `${baseUrl}/about/press`, children: 'Press office' })}</li>
+            <li>{renderLink({ href: `${baseUrl}/invest`, children: 'Peer-to-peer investments' })}</li>
           </List>
         </FlexCol>
         <FlexCol xs={12} s={6} l={3} className="mb-8 s:mb-9">
@@ -85,7 +85,10 @@ const ZopaFooter = ({
             </li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/modern-slavery`, children: 'Modern slavery' })}</li>
             <li className="mb-4">{renderLink({ href: `${baseUrl}/terms-of-use`, children: 'Terms of use' })}</li>
-            <li>{renderLink({ href: `${baseUrl}/investor-principles`, children: 'Investor principles' })}</li>
+            <li className="mb-4">
+              {renderLink({ href: `${baseUrl}/investor-principles`, children: 'Investor principles' })}
+            </li>
+            <li>{renderLink({ href: `${baseUrl}/remuneration`, children: 'Remuneration' })}</li>
           </List>
         </FlexCol>
         <FlexCol xs={12} s={6} l={3} className="mb-8 s:mb-9">