+++ src/components/organisms/Card/Card/Card.tsx
@@ -1,13 +1,12 @@
 import styled, { css } from 'styled-components';
-import { typography } from '../../../../constants/typography';
 import { colors } from '../../../../constants/colors';
+import { typography } from '../../../../constants/typography';
 import { CardImageContainer } from '../CardImage/CardImage';
-import CardHeading from '../CardHeading/CardHeading';
 import CardText from '../CardText/CardText';
 import CardLineItem from '../CardLineItem/CardLineItem';
 
 export type CardLayout = 'horizontal' | 'vertical';
-export type CardStyling = 'primary' | 'secondary' | 'brand' | 'action';
+export type CardStyling = 'primary' | 'secondary' | 'brand' | 'action' | 'info';
 
 export interface CardProps {
   /**
@@ -37,6 +36,7 @@ const headingSizes: Record<CardStyling, string> = {
   primary: typography.sizes.heading.h5,
   brand: typography.sizes.heading.h6,
   action: typography.sizes.heading.h6,
+  info: typography.sizes.heading.h6,
 };
 
 const textSizes: Record<CardStyling, string> = {
@@ -44,6 +44,7 @@ const textSizes: Record<CardStyling, string> = {
   primary: typography.sizes.text.body,
   brand: typography.sizes.text.small,
   action: typography.sizes.text.small,
+  info: typography.sizes.text.small,
 };
 
 const boxShadowStyle: Record<CardStyling, string> = {
@@ -51,6 +52,7 @@ const boxShadowStyle: Record<CardStyling, string> = {
   secondary: `0 1px 0 0 ${colors.greyLight}`,
   brand: `none`,
   action: `none`,
+  info: `none`,
 };
 
 const borderStyle: Record<CardStyling, string> = {
@@ -58,6 +60,7 @@ const borderStyle: Record<CardStyling, string> = {
   secondary: `1px solid ${colors.greyLighter}`,
   brand: `1px solid ${colors.brand}`,
   action: `1px solid ${colors.greyLighter}`,
+  info: `1px solid ${colors.greyLightest}`,
 };
 
 const backgroundStyle: Record<CardStyling, string> = {
@@ -65,6 +68,7 @@ const backgroundStyle: Record<CardStyling, string> = {
   secondary: `${colors.white};`,
   brand: `${colors.brandLight};`,
   action: `${colors.white};`,
+  info: `${colors.greyLightest}`,
 };
 
 const Card = styled.div<CardProps>`
@@ -84,7 +88,7 @@ const Card = styled.div<CardProps>`
           ${layoutImageStyles[layout]}
         `};
   }
-  & ${CardHeading} {
+  & .zrc__card-heading {
     font-size: ${({ styling = 'secondary' }) =>
       css`
         ${headingSizes[styling]}
@@ -95,9 +99,6 @@ const Card = styled.div<CardProps>`
       css`
         ${textSizes[styling]}
       `};
-  }
-  & > ${CardLineItem} {
-    align-self: center;
   }
   ${({ styling }) =>
     styling === 'action'